backend/requirements.txt
# AfterIDE Backend Requirements
# Note: Pip is automatically upgraded to latest version in setup.sh to eliminate warning messages
# This ensures users never see "pip version X is available" warnings

# Latest pip (automatically upgraded during setup)
pip>=25.0.0

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and caching
redis==5.0.1
aioredis==2.0.1

# Security and validation
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0

# Container and execution
docker==6.1.3
psutil==5.9.6

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
python-json-logger==2.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 